backend/pyproject.toml
[project]
name = "chemplan-backend"
version = "0.1.0"
description = "Backend for chemplan: FastAPI wrapping AiZynthFinder"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "aizynthfinder>=4.3",
    "anthropic>=0.92",
    "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
